pyproject.toml
[project]
name = "convert-images-web"
version = "0.1.0"
description = "..."
authors = [
    { name = "Jürgen Schoenemeyer", email = "js007@web.de" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "pillow>=11.0.0",
    "pillow-avif-plugin>=1.4.6",
]
[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "types-python-dateutil>=2.9.0.20241206",
]

[tool.basedpyright]
deprecateTypingAliases =  false
reportAny = false
reportImplicitRelativeImport = false
reportUnusedCallResult = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportExplicitAny = false
